+++ biotrack/src/main/java/com/ProgWebII/biotrack/dto/response/BuscarUsuarioPorIdResponse.java
@@ -1,4 +1,4 @@
-package com.ProgWebII.biotrack.dto;
+package com.ProgWebII.biotrack.dto.response;
 
 import java.time.LocalDate;
