+++ biotrack/src/main/java/com/ProgWebII/biotrack/dto/response/UsuarioComUltimaMedidaResponse.java
@@ -1,13 +1,12 @@
-package com.ProgWebII.biotrack.dto;
+package com.ProgWebII.biotrack.dto.response;
 
 import java.time.LocalDate;
-import java.util.List;
 
-public record UsuarioResponse(
+public record UsuarioComUltimaMedidaResponse(
         Long id,
         String name,
         LocalDate birthDate,
         String zipCode,
         String email,
-        List<MedidaResponse> medidas
+        MedidaResponse ultimaMedida
 ) {}
