+++ biotrack/src/main/java/com/ProgWebII/biotrack/dto/response/UsuarioSemMedidasResponse.java
@@ -1,8 +1,8 @@
-package com.ProgWebII.biotrack.dto;
+package com.ProgWebII.biotrack.dto.response;
 
 import java.time.LocalDate;
 
-public record ListarTodosUsuariosResponse(
+public record UsuarioSemMedidasResponse(
         Long id,
         String name,
         LocalDate birthDate,
