+++ biotrack/src/main/java/com/ProgWebII/biotrack/service/UserService.java
@@ -1,7 +1,7 @@
 package com.ProgWebII.biotrack.service;
 
-import com.ProgWebII.biotrack.dto.*;
 import com.ProgWebII.biotrack.dto.request.UserRequest;
+import com.ProgWebII.biotrack.dto.response.*;
 import com.ProgWebII.biotrack.model.Measure;
 import com.ProgWebII.biotrack.model.User;
 import com.ProgWebII.biotrack.repository.UserRepository;
